@Zarel 2018-03-05
I wonder if caching relevant effect objects every time effects are added/removed from the environment would represent a perf win in general.