@strager 2018-03-05
> I wonder if caching relevant effect objects every time effects are added/removed from the environment would represent a perf win in general.

That's what I was going to do originally, but I wasn't sure if such changes would be accepted, and I found some easier low-hanging fruit anyway.